moved more of the helpers function from routs

+++ client/src/components/Register.js
@@ -102,6 +102,10 @@ function Register({ intialSkierList, updateLoggedIn, getName, getUserId }) {
     e.preventDefault()
     try {
       const resp = await axios.post('/users/register', userData);
+      if(resp.data.message === "user is already registered"){
+        setShowAlert(true)
+        return
+      }
       localStorage.setItem("skiBuddyToken", resp.data.token);
       updateLoggedIn(true);
       getName(resp.data.name);
